impr(repository): implementation following clean architecture
- Added SQLite repositories for `Restaurant` and `Review`
- Included adapters for handling repository output
- Fixed `ReviewFactory` bug
