+++ src/main/java/entity/GeminiReviewFactory.java
@@ -1,12 +1,9 @@
 package entity;
 
-/**
- * Factory class for creating Review instances that are summarized by Gemini API.
- */
-public class GeminiReviewFactory implements ReviewFactory {
+public class GeminiReviewFactory implements ReviewFactoryWithoutAuthor {
 
     @Override
-    public Review createReview(String restaurantId, String author, String content) {
+    public Review createReview(String restaurantId, String content) {
         return new Review(restaurantId, "Gemini", content, true);
     }
 }
