+++ src/main/java/entity/ReviewFactoryWithAuthor.java
@@ -1,0 +1,5 @@
+package entity;
+
+public interface ReviewFactoryWithAuthor {
+    Review createReview(String restaurantId, String author, String content);
+}
