+++ src/main/java/entity/ReviewFactoryWithoutAuthor.java
@@ -1,0 +1,5 @@
+package entity;
+
+public interface ReviewFactoryWithoutAuthor {
+    Review createReview(String restaurantId, String content);
+}
