+++ src/main/java/entity/UserReviewFactory.java
@@ -3,7 +3,7 @@ package entity;
 /**
  * Factory class for creating Review instances that are collected from Google Places API.
  */
-public class UserReviewFactory implements ReviewFactory {
+public class UserReviewFactory implements ReviewFactoryWithAuthor {
 
     @Override
     public Review createReview(String restaurantId, String author, String content) {
