+++ src/main/java/interface_adapter/data/RestaurantDataAdapter.java
@@ -1,0 +1,10 @@
+package interface_adapter.data;
+
+import entity.Restaurant;
+
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
+public interface RestaurantDataAdapter {
+     Restaurant adaptToRestaurant(ResultSet rs) throws SQLException;
+}
