+++ src/main/java/interface_adapter/data/ReviewDataAdapter.java
@@ -1,0 +1,9 @@
+package interface_adapter.data;
+
+import entity.Review;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
+public interface ReviewDataAdapter {
+    Review adaptToReview(ResultSet rs) throws SQLException;
+}
